Auth bug fix 2.0

+++ src/main/java/com/adminmodule/controller/EmployeeController.java
@@ -1,6 +1,7 @@
 package com.adminmodule.controller;
 
 import com.adminmodule.security.AuthService;
+import com.adminmodule.security.JWTUtility;
 import com.adminmodule.service.EmployeeService;
 import com.adminmodule.service.dto.EmployeeDTO;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -12,12 +13,12 @@ import org.springframework.web.bind.annotation.*;
 public class EmployeeController {
     private final EmployeeService employeeService;
 
-    private final AuthService authService;
+    private final JWTUtility jwtUtility;
 
     @Autowired
-    public EmployeeController(EmployeeService employeeService, AuthService authService) {
+    public EmployeeController(EmployeeService employeeService, JWTUtility jwtUtility) {
         this.employeeService = employeeService;
-        this.authService = authService;
+        this.jwtUtility = jwtUtility;
     }
 
     @GetMapping
@@ -25,7 +26,13 @@ public class EmployeeController {
         return ResponseEntity.ok(employeeService.getAllEmployees());
     }
     @GetMapping("/{id}")
-    public ResponseEntity<?> getEmployeeById(@PathVariable("id") Long id) {
+    public ResponseEntity<?> getEmployeeById(@PathVariable("id") Long id, @RequestHeader("Authorization") String authorizationHeader) {
+        String jwtToken = authorizationHeader.substring(7);
+        String username = jwtUtility.getUsernameFromToken(jwtToken);
+        EmployeeDTO employeeDTO = employeeService.getEmployeeByUsername(username);
+        if (employeeDTO.getId() != id) {
+            return ResponseEntity.status(403).build();
+        }
         return ResponseEntity.ok(employeeService.getEmployeeById(id));
     }
     @PostMapping
@@ -33,11 +40,25 @@ public class EmployeeController {
         return ResponseEntity.ok(employeeService.addEmployee(employeeDTO));
     }
     @PutMapping("/{id}")
-    public ResponseEntity<?> updateEmployee(@PathVariable("id") Long id, @RequestBody EmployeeDTO employeeDTO) {
+    public ResponseEntity<?> updateEmployee(@PathVariable("id") Long id, @RequestBody EmployeeDTO employeeDTO,
+                                            @RequestHeader("Authorization") String authorizationHeader) {
+        String jwtToken = authorizationHeader.substring(7);
+        String username = jwtUtility.getUsernameFromToken(jwtToken);
+        EmployeeDTO employeeDTO1 = employeeService.getEmployeeByUsername(username);
+        if (employeeDTO1.getId() != id) {
+            return ResponseEntity.status(403).build();
+        }
         return ResponseEntity.ok(employeeService.updateEmployee(id, employeeDTO));
     }
     @DeleteMapping("/{id}")
-    public ResponseEntity<?> deleteEmployee(@PathVariable("id") Long id) {
+    public ResponseEntity<?> deleteEmployee(@PathVariable("id") Long id,
+                                            @RequestHeader("Authorization") String authorizationHeader) {
+        String jwtToken = authorizationHeader.substring(7);
+        String username = jwtUtility.getUsernameFromToken(jwtToken);
+        EmployeeDTO employeeDTO1 = employeeService.getEmployeeByUsername(username);
+        if (employeeDTO1.getId() != id) {
+            return ResponseEntity.status(403).build();
+        }
         employeeService.deleteEmployee(id);
         return ResponseEntity.ok().build();
     }
