+++ src/main/java/com/adminmodule/controller/VendorAdminController.java
@@ -1,11 +1,14 @@
 package com.adminmodule.controller;
 
 import com.adminmodule.security.AuthService;
+import com.adminmodule.security.JWTUtility;
 import com.adminmodule.service.VendorAdminService;
 import com.adminmodule.service.dto.VendorAdminDTO;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+
+import java.util.List;
 
 @RestController
 @RequestMapping("/api/v1/vendor-admin")
@@ -13,12 +16,12 @@ public class VendorAdminController {
 
     private final VendorAdminService vendorAdminService;
 
-    private final AuthService authService;
+    private final JWTUtility jwtUtility;
 
     @Autowired
-    public VendorAdminController(VendorAdminService vendorAdminService, AuthService authService) {
+    public VendorAdminController(VendorAdminService vendorAdminService, JWTUtility jwtUtility) {
         this.vendorAdminService = vendorAdminService;
-        this.authService = authService;
+        this.jwtUtility = jwtUtility;
     }
 
     @GetMapping("all")
@@ -26,7 +29,15 @@ public class VendorAdminController {
         return ResponseEntity.ok(vendorAdminService.getAllVendorAdmins());
     }
     @GetMapping("/vendor/{id}")
-    public ResponseEntity<?> getVendorAdminsByVendorId(@PathVariable Long id) {
+    public ResponseEntity<?> getVendorAdminsByVendorId(@PathVariable Long id,
+                                                       @RequestHeader("Authorization") String authorizationHeader) {
+        String jwtToken = authorizationHeader.substring(7);
+        String username = jwtUtility.getUsernameFromToken(jwtToken);
+        VendorAdminDTO vendorAdminDTO = vendorAdminService.getVendorAdminByUsername(username);
+        List<String> roles = jwtUtility.getRoleFromToken(jwtToken);
+        if (vendorAdminDTO.getId() != id || !roles.contains("ROLE_ADMIN")) {
+            return ResponseEntity.status(403).build();
+        }
         return ResponseEntity.ok(vendorAdminService.getVendorAdminById(id));
     }
 
@@ -35,13 +46,28 @@ public class VendorAdminController {
         return ResponseEntity.ok(vendorAdminService.addVendorAdmin(vendorAdminDTO));
     }
     @PutMapping("/{id}")
-    public ResponseEntity<?> updateVendorAdmin(@PathVariable Long id, @RequestBody VendorAdminDTO vendorAdminDTO) {
+    public ResponseEntity<?> updateVendorAdmin(@PathVariable Long id, @RequestBody VendorAdminDTO vendorAdminDTO,
+                                               @RequestHeader("Authorization") String authorizationHeader) {
+        String jwtToken = authorizationHeader.substring(7);
+        String username = jwtUtility.getUsernameFromToken(jwtToken);
+        VendorAdminDTO vendorAdminDTO1 = vendorAdminService.getVendorAdminByUsername(username);
+        List<String> roles = jwtUtility.getRoleFromToken(jwtToken);
+        if (vendorAdminDTO1.getId() != id || !roles.contains("ROLE_ADMIN")) {
+            return ResponseEntity.status(403).build();
+        }
         return ResponseEntity.ok(vendorAdminService.updateVendorAdmin(id, vendorAdminDTO));
     }
     @DeleteMapping("/{id}")
-    public ResponseEntity<?> deleteVendorAdmin(@PathVariable Long id) {
+    public ResponseEntity<?> deleteVendorAdmin(@PathVariable Long id,
+                                               @RequestHeader("Authorization") String authorizationHeader) {
+        String jwtToken = authorizationHeader.substring(7);
+        String username = jwtUtility.getUsernameFromToken(jwtToken);
+        VendorAdminDTO vendorAdminDTO1 = vendorAdminService.getVendorAdminByUsername(username);
+        List<String> roles = jwtUtility.getRoleFromToken(jwtToken);
+        if (vendorAdminDTO1.getId() != id || !roles.contains("ROLE_ADMIN")) {
+            return ResponseEntity.status(403).build();
+        }
         vendorAdminService.deleteVendorAdmin(id);
         return ResponseEntity.ok().build();
     }
-
 }
